+++ srcs/HttpManager.cpp
@@ -21,7 +21,7 @@
 #include "HttpManager.hpp"
 
 HttpManager::HttpManager(const ConfigParser::ServerConfig& config)
-    : _config(config), _parser() {}
+    : _config(config), _parser() {(void) _config;}
 
 std::string HttpManager::processHttpRequest(const std::string& raw_request,
                                             HttpRequest& request,
